requirements.txt
# ========== NL2SQL BI Agent 项目依赖 ==========
# Web 框架
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
python-dotenv==1.0.1
sse-starlette==2.1.0

# ========== LLM / Agent 框架 ==========
langchain==0.3.10
langchain-core==0.3.23
langchain-community==0.3.10
langchain-openai==0.2.9
langchain-ollama==0.2.0
langgraph==0.2.58

# ========== 向量数据库（Milvus） ==========
pymilvus==2.4.9
milvus-lite==2.4.9

# ========== Embedding / 向量化 ==========
sentence-transformers==3.0.1
numpy==1.26.4

# ========== 数据库连接 ==========
# pymysql==1.1.1  # MySQL 驱动，学习环境不需要
psycopg[binary]==3.2.3
sqlalchemy==2.0.35
# sqlite3（Python 内置，无需安装）

# ========== 数据分析与可视化推荐 ==========
pandas==2.2.3

# ========== 工具库 ==========
httpx==0.27.2
aiofiles==24.1.0
psutil==6.0.0
rank-bm25==0.2.2
beautifulsoup4==4.12.3
lxml==5.3.0

# ========== 前端（Node.js 项目） ==========
# vue@3, vite@5, echarts@5, typescript 等通过 npm install
